erase Audio and OSC Input nodes while editing inputs

+++ farq2014/src/ParameterInputGen/DTAudioMap.h
@@ -24,8 +24,8 @@ public:
     int     band;
     float   inputMinValue;
     float   inputMaxValue;
-    int     paramMinValue;
-    int     paramMaxValue;
+    float   paramMinValue;
+    float   paramMaxValue;
     
 };
 #endif /* defined(__ofApp__DTAudioMap__) */
